# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: HyperMind remote API
  version: 3.0.0
  description: >-
    Frozen NCPR v3. Mandatory mutual TLS and a capability bearer token.
    Actor and admin listeners are separate. The existing fourteen MCP verb
    names are preserved; arguments and tool envelopes have identical semantics.
    Crypto-shred and projection rebuild require the admin listener and token.
    Logical fade and retraction remain actor-authorized operations.
security:
  - capability: []
    mutualTLS: []
paths:
...
  /v1/admin/{verb}:
    post:
      parameters:
        - name: verb
          in: path
          required: true
          schema:
            type: string
            enum: [health, stats, verify, rebuild, forget]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCall'
      responses:
        '200': *envelope
        '400': *envelope
        '401': *envelope
        '403': *envelope
        '429': *envelope
        '503': *envelope
components:
  securitySchemes:
    capability:
      type: http
      scheme: bearer
      description: 32-byte capability token as 64 hexadecimal characters
    mutualTLS:
      type: mutualTLS
  schemas:
...
    AdminCall:
      type: object
      required: [connection_id, request_id]
      properties:
        connection_id:
          type: string
          pattern: '^[0-9a-fA-F]{32}$'
        request_id:
          type: integer
          format: uint64
          minimum: 1
        actor:
          type: integer
          format: uint16
          description: Required nonzero actor for every operation except health
        arguments:
          type: object
          description: Rebuild requires name; forget requires action=crypto_shred
          additionalProperties: true
...